requirements.txt
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
db-dtypes==1.4.3
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
idna==3.10
isodate==0.7.2
lxml==5.4.0
more-itertools==10.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9.1
simple-salesforce==1.12.6
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
zeep==4.3.1
